# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shortlist - AI Resume Screening API",
    "description": "\n**Shortlist** screens resumes against a job description and ranks candidates by fit.\n\n### How it works\n1. **Create a job** - paste a description; required / preferred skills, minimum experience and education are extracted automatically (and can be edited).\n2. **Upload resumes** (PDF / DOCX, scanned PDFs via OCR) - each is parsed into name, e-mail, phone, skills, education, experience and certifications.\n3. **Get a ranking** - every candidate receives a **0-100 match score** with a transparent breakdown\n   (skills 50 %, semantic similarity 20 %, experience 20 %, education 10 %) plus the matched, related and missing skills.\n4. **Filter, shortlist, export** to Excel / CSV, e-mail shortlisted candidates and generate tailored interview questions.\n\nErrors always use the shape `{\"error\": {\"code\": \"...\", \"message\": \"...\"}}`.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/skills/top": {
      "get": {
        "tags": [
          "Meta"
        ],
        "summary": "Most common candidate skills (powers filter suggestions)",
        "operationId": "top_skills_api_skills_top_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 40,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/parse-description": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Extract required/preferred skills, experience and education from a job description",
        "operationId": "parse_description_api_jobs_parse_description_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobParseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "List jobs with screening statistics",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "type": "array",
                  "title": "Response List Jobs Api Jobs Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Create a job (requirements are auto-extracted unless you provide them) and score all candidates",
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "put": {
        "tags": [
          "Jobs"
        ],
        "summary": "Edit a job and re-score all candidates",
        "operationId": "update_job_api_jobs__job_id__put",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Delete Job",
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/resumes/upload": {
      "post": {
        "tags": [
          "Resumes"
        ],
        "summary": "Upload one or many resumes (PDF / DOCX). Each is parsed, stored and scored against every job.",
        "operationId": "upload_resumes_api_resumes_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resumes_api_resumes_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "List / filter all candidates",
        "operationId": "list_candidates_api_candidates_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search name, e-mail or headline",
              "title": "Q"
            },
            "description": "Search name, e-mail or headline"
          },
          {
            "name": "skills",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated skills, e.g. `python,aws`",
              "title": "Skills"
            },
            "description": "Comma-separated skills, e.g. `python,aws`"
          },
          {
            "name": "skill_mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "any"
              ],
              "type": "string",
              "description": "Candidate must have ALL or ANY of `skills`",
              "default": "all",
              "title": "Skill Mode"
            },
            "description": "Candidate must have ALL or ANY of `skills`"
          },
          {
            "name": "min_experience",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum years of experience",
              "title": "Min Experience"
            },
            "description": "Minimum years of experience"
          },
          {
            "name": "max_experience",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Experience"
            }
          },
          {
            "name": "education",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "diploma",
                    "bachelor",
                    "master",
                    "phd"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum education level",
              "title": "Education"
            },
            "description": "Minimum education level"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new",
                    "shortlisted",
                    "rejected"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateBrief"
                  },
                  "title": "Response List Candidates Api Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{cid}": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "Get Candidate",
        "operationId": "get_candidate_api_candidates__cid__get",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateFull"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "delete": {
        "tags": [
          "Candidates"
        ],
        "summary": "Delete Candidate",
        "operationId": "delete_candidate_api_candidates__cid__delete",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobParseRequest": {
        "properties": {
          "description": {
            "type": "string",
            "minLength": 20,
            "title": "Description",
            "description": "Free-text job description"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "JobParseRequest"
      },
...
      "UploadResponse": {
        "properties": {
          "processed": {
            "type": "integer",
            "title": "Processed"
          },
          "duplicates": {
            "type": "integer",
            "title": "Duplicates"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/UploadItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "processed",
          "duplicates",
          "failed",
          "results"
        ],
        "title": "UploadResponse"
      },
...
    }
  },
...
}